Refactor ipywidget tests to use real kernel to wait for idle
